+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - EMPLOYEE : set seniorityDate by hireDate
 - Studio : Add CSRF protection for every request header.
 - Add CSRF protection for timesheet line editor and project planning editor
+- Studio : Add support of menu creation for custom model and BPM node.
 
 ## Improvements
 - USER : Default User language is based on application.locale from application.properties